pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-claw"
version = "0.1.0"
description = "A lightweight AI agent framework in ~5K lines of code"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nano Agent Contributors"}
]
keywords = [
    "ai",
    "agent",
    "llm",
    "mcp",
    "openai",
    "anthropic",
    "google",
    "gemini",
    "assistant",
    "chatbot",
    "automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "aiohttp>=3.9.0",
    "mcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/nano-claw"
Documentation = "https://github.com/yourusername/nano-claw/tree/main/docs"
Repository = "https://github.com/yourusername/nano-claw"
"Bug Tracker" = "https://github.com/yourusername/nano-claw/issues"

[project.scripts]
nano-claw = "nano_claw.main:main"
nano-claw-cli = "nano_claw.cli:main"

[tool.setuptools]
packages = ["nano_claw"]

[tool.setuptools.package-data]
nano_claw = ["*.yaml", "*.md"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
